Fixed compilation issues with FC3 on an Athlon 64.

+++ src/protocol.cc
@@ -156,7 +156,7 @@ bool Packet::DecodeUInt32(unsigned int &i)
   return false;
 }
 
-bool Packet::DecodeUInt64(unsigned long long  &i)
+bool Packet::DecodeUInt64(guint64  &i)
 {
   char *b = rxBuffer->getBuffer();
 
@@ -261,7 +261,7 @@ bool Packet::EncodeUInt32(unsigned int i)
   return true;
 }
 
-bool Packet::EncodeUInt64(unsigned long long  i)
+bool Packet::EncodeUInt64(guint64  i)
 {
 
   txBuffer->putc(i2a(eGPSIM_TYPE_UINT64 /16));
